requirements.txt
altair==4.2.2
attrs==25.1.0
audioread==3.0.1
blinker==1.9.0
CacheControl==0.14.2
cachetools==4.2.4
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
cryptography==44.0.0
decorator==5.1.1
entrypoints==0.4
et_xmlfile==2.0.0
filelock==3.17.0
firebase-admin==5.0.0
Flask==3.0.2
Flask-Cors==5.0.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==1.34.1
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.11.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.48.2
httplib2==0.22.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.6.1
itemadapter==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.25.0
protobuf==3.20.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
Pympler==1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
scrapeops_scrapy==0.5.6
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
soundfile==0.13.0
soxr==0.5.0.post1
standard-imghdr==3.13.0
streamlit==1.41.1
streamlit-cropper==0.2.2
streamlit-image-coordinates==0.2.0
streamlit-image-select==0.6.0
tenacity==8.5.0
threadpoolctl==3.5.0
tld==0.13
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
validators==0.34.0
Werkzeug==3.1.3
zipp==3.21.0